refactor: moved bookmark_manager under api for organization
